add user check

+++ app/Http/Controllers/ListingController.php
@@ -55,6 +55,10 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $form_fields = $request->validate(
             [
                 'title'    => 'required',
@@ -78,6 +82,10 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully!');
     }
